skillzone/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.0
django-cors-headers>=4.2.0
psycopg2-binary>=2.9.6  # For PostgreSQL support
python-dotenv>=1.0.0
